Ball display updated

+++ src/main/sensor/response/BallsResponse.kt
@@ -1,5 +1,4 @@
 package main.sensor.response
 
 data class BallsResponse(val data:List<Ball>): BaseResponse(null, null)
-
-data class Ball(val x: Int, val y: Int)
+data class Ball(val x: Float, val y: Float)
